# delta-agent/docs/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Delta Agent API",
    "version": "1.0.0",
    "description": "Deterministic execution API contract for session-scoped operations, DeltaShot replayability, workflow orchestration, and artifact versioning."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/v1/workspaces": {
      "post": {
        "summary": "Create workspace",
        "operationId": "createWorkspace",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceCreateResponseEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}": {
      "get": {
        "summary": "Get workspace",
        "operationId": "getWorkspace",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspaceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace view",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceViewEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/sessions": {
      "get": {
        "summary": "List sessions in workspace",
        "operationId": "listWorkspaceSessions",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspaceId"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "post": {
        "summary": "Create session",
        "operationId": "createSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreateResponseEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/messages": {
      "post": {
        "summary": "Send message (primary deterministic entrypoint)",
        "operationId": "sendMessage",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequestV1"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deterministic message delta response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMessageResponseEnvelope"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "get": {
        "summary": "Get session messages",
        "operationId": "getSessionMessages",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/state": {
      "get": {
        "summary": "Get session state",
        "operationId": "getSessionState",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Session state snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateResponseEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/deltashots": {
      "get": {
        "summary": "List DeltaShots",
        "operationId": "listDeltashots",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "DeltaShot list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltashotListEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deltashots/{deltashotId}": {
      "get": {
        "summary": "Get DeltaShot",
        "operationId": "getDeltashot",
        "parameters": [
          {
            "$ref": "#/components/parameters/deltashotId"
          }
        ],
        "responses": {
          "200": {
            "description": "DeltaShot detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltashotViewEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/rollback": {
      "post": {
        "summary": "Rollback session to target DeltaShot",
        "operationId": "rollbackSession",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rollback result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponseEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/artifacts": {
      "post": {
        "summary": "Create or update artifact",
        "operationId": "writeArtifact",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Artifact delta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactEnvelopeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifactId}": {
      "get": {
        "summary": "Get artifact metadata",
        "operationId": "getArtifact",
        "parameters": [
          {
            "$ref": "#/components/parameters/artifactId"
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactViewEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/artifacts/{artifactId}/versions": {
      "get": {
        "summary": "Get artifact versions",
        "operationId": "getArtifactVersions",
        "parameters": [
          {
            "$ref": "#/components/parameters/artifactId"
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact versions list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionListEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifactId}/versions/{version}": {
      "get": {
        "summary": "Get specific artifact version",
        "operationId": "getArtifactVersion",
        "parameters": [
          {
            "$ref": "#/components/parameters/artifactId"
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact version content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionViewEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/workflows/start": {
      "post": {
        "summary": "Start workflow",
        "operationId": "startWorkflow",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workflow state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStateEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/execute-next": {
      "post": {
        "summary": "Execute next queued workflow step",
        "operationId": "executeNextWorkflowJob",
        "responses": {
          "200": {
            "description": "Workflow execution result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowExecutionResultEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/workflows/{workflowId}/state": {
      "get": {
        "summary": "Get workflow state",
        "operationId": "getWorkflowState",
        "parameters": [
          {
            "$ref": "#/components/parameters/workflowId"
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStateEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/workflows/{workflowId}/step": {
      "post": {
        "summary": "Advance workflow step",
        "operationId": "advanceWorkflowStep",
        "parameters": [
          {
            "$ref": "#/components/parameters/workflowId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStepRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workflow state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStateEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/agent": {
      "post": {
        "summary": "Force agent selection",
        "operationId": "forceAgentSelection",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForceAgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Forced agent result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSelectionViewEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/agents/logs": {
      "get": {
        "summary": "Get agent execution logs",
        "operationId": "getAgentLogs",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent log entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentLogListEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/trace": {
      "get": {
        "summary": "Get execution trace",
        "operationId": "getTrace",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Execution trace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionTraceEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "workspaceId": {
        "name": "workspaceId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "sessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "deltashotId": {
        "name": "deltashotId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "artifactId": {
        "name": "artifactId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "workflowId": {
        "name": "workflowId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "ApiError": {
        "description": "Standard API error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateSessionRequest": {
        "type": "object",
        "required": [
          "workspace_id",
          "name"
        ],
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "SendMessageRequestV1": {
        "type": "object",
        "required": [
          "content",
          "mode"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "chat",
              "workflow",
              "execution"
            ]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "RollbackRequest": {
        "type": "object",
        "required": [
          "target_deltashot_id",
          "mode"
        ],
        "properties": {
          "target_deltashot_id": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "hard",
              "soft"
            ]
          }
        }
      },
      "ArtifactWriteRequest": {
        "type": "object",
        "required": [
          "session_id",
          "type",
          "content"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "WorkspaceCreateResponseEnvelope": {
        "type": "object",
        "required": [
          "workspace_id",
          "created_at",
          "meta"
        ],
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "created_at": {
            "type": "integer"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "WorkspaceViewEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/WorkspaceView"
          },
          {
            "type": "object",
            "required": [
              "meta"
            ],
            "properties": {
              "meta": {
                "$ref": "#/components/schemas/ApiMeta"
              }
            }
          }
        ]
      },
...
      "SessionCreateResponseEnvelope": {
        "type": "object",
        "required": [
          "session_id",
          "created_at",
          "meta"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "created_at": {
            "type": "integer"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "SessionListEnvelope": {
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionView"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "SendMessageResponseEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SendMessageResponseV1"
          },
          {
            "type": "object",
            "required": [
              "meta"
            ],
            "properties": {
              "meta": {
                "$ref": "#/components/schemas/ApiMeta"
              }
            }
          }
        ]
      },
...
      "SessionStateResponseEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SessionStateResponse"
          },
          {
            "type": "object",
            "required": [
              "meta"
            ],
            "properties": {
              "meta": {
                "$ref": "#/components/schemas/ApiMeta"
              }
            }
          }
        ]
      },
...
      "MessageListEnvelope": {
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageRecord"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "DeltashotViewEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DeltashotView"
          },
          {
            "type": "object",
            "required": [
              "meta"
            ],
            "properties": {
              "meta": {
                "$ref": "#/components/schemas/ApiMeta"
              }
            }
          }
        ]
      },
      "DeltashotListEnvelope": {
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeltashotView"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "RollbackResponseEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RollbackResponse"
          },
          {
            "type": "object",
            "required": [
              "meta"
            ],
            "properties": {
              "meta": {
                "$ref": "#/components/schemas/ApiMeta"
              }
            }
          }
        ]
      },
...
      "ArtifactEnvelopeResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArtifactEnvelope"
          },
          {
            "type": "object",
            "required": [
              "meta"
            ],
            "properties": {
              "meta": {
                "$ref": "#/components/schemas/ApiMeta"
              }
            }
          }
        ]
      },
      "ArtifactViewEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArtifactView"
          },
          {
            "type": "object",
            "required": [
              "meta"
            ],
            "properties": {
              "meta": {
                "$ref": "#/components/schemas/ApiMeta"
              }
            }
          }
        ]
      },
...
    }
  }
}